refactor : 부모 패키지 org 생성
