+++ src/main/java/org/CalculatorApplication.java
@@ -1,3 +1,5 @@
+package org;
+
 public class CalculatorApplication {
 
     public static void main(String[] args) {
