--- src/main/java/service/CalculationService.java
@@ -1,6 +1,0 @@
-package service;
-
-public class CalculationService {
-
-    //TODO : 계산 비즈니스 로직
-}
